remove sudo -u www-data (works as root)

+++ debian/owncloud_install.sh
@@ -62,7 +62,10 @@ echo -e "\e[0m"
 
 # Install MYSQL 5.7
 aptitude install debconf-utils -y
+if [ -d /var/lib/mysql ];
+then
 rm -R /var/lib/mysql
+fi
 echo "deb http://repo.mysql.com/apt/debian/ jessie mysql-$MYSQL_VERSION" > /etc/apt/sources.list.d/mysql.list
 echo "deb-src http://repo.mysql.com/apt/debian/ jessie mysql-$MYSQL_VERSION" >> /etc/apt/sources.list.d/mysql.list
 echo mysql-community-server mysql-community-server/root-pass password $MYSQL_PASS | debconf-set-selections
@@ -150,11 +153,12 @@ bash $SCRIPTS/setup_secure_permissions_owncloud.sh
 
 # Install ownCloud
 cd $OCPATH
-su www-data
+
 php occ maintenance:install --database "mysql" --database-name "owncloud_db" --database-user "root" --database-pass "$MYSQL_PASS" --admin-user "ocadmin" --admin-pass "owncloud"
 echo
 echo ownCloud version:
-sudo -u www-data php $OCPATH/occ status
+
+php $OCPATH/occ status
 echo
 sleep 3
 
@@ -228,18 +232,19 @@ service apache2 restart
 
 ## Set config values
 # Experimental apps
-sudo -u www-data php $OCPATH/occ config:system:set appstore.experimental.enabled --value="true"
+
+php $OCPATH/occ config:system:set appstore.experimental.enabled --value="true"
 # Default mail server (make this user configurable?)
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpmode --value="smtp"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpauth --value="1"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpport --value="465"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtphost --value="smtp.gmail.com"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpauthtype --value="LOGIN"
-sudo -u www-data php $OCPATH/occ config:system:set mail_from_address --value="www.en0ch.se"
-sudo -u www-data php $OCPATH/occ config:system:set mail_domain --value="gmail.com"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpsecure --value="ssl"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtpname --value="www.en0ch.se@gmail.com"
-sudo -u www-data php $OCPATH/occ config:system:set mail_smtppassword --value="techandme_se"
+php $OCPATH/occ config:system:set mail_smtpmode --value="smtp"
+php $OCPATH/occ config:system:set mail_smtpauth --value="1"
+php $OCPATH/occ config:system:set mail_smtpport --value="465"
+php $OCPATH/occ config:system:set mail_smtphost --value="smtp.gmail.com"
+php $OCPATH/occ config:system:set mail_smtpauthtype --value="LOGIN"
+php $OCPATH/occ config:system:set mail_from_address --value="www.en0ch.se"
+php $OCPATH/occ config:system:set mail_domain --value="gmail.com"
+php $OCPATH/occ config:system:set mail_smtpsecure --value="ssl"
+php $OCPATH/occ config:system:set mail_smtpname --value="www.en0ch.se@gmail.com"
+php $OCPATH/occ config:system:set mail_smtppassword --value="techandme_se"
 
 # Install Libreoffice Writer to be able to read MS documents.
 echo -ne '\n' | sudo apt-add-repository ppa:libreoffice/libreoffice-4-4
@@ -259,8 +264,9 @@ fi
 
 # Enable documents
 if [ -d $OCPATH/apps/documents ]; then
-sudo -u www-data php $OCPATH/occ app:enable documents
-sudo -u www-data php $OCPATH/occ config:system:set preview_libreoffice_path --value="/usr/bin/libreoffice"
+
+php $OCPATH/occ app:enable documents
+php $OCPATH/occ config:system:set preview_libreoffice_path --value="/usr/bin/libreoffice"
 fi
 
 # Download and install Contacts
@@ -276,7 +282,7 @@ fi
 
 # Enable Contacts
 if [ -d $OCPATH/apps/contacts ]; then
-sudo -u www-data php $OCPATH/occ app:enable contacts
+php $OCPATH/occ app:enable contacts
 fi
 
 # Download and install Calendar
@@ -292,7 +298,7 @@ fi
 
 # Enable Calendar
 if [ -d $OCPATH/apps/calendar ]; then
-sudo -u www-data php $OCPATH/occ app:enable calendar
+php $OCPATH/occ app:enable calendar
 fi
